backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
PyJWT==2.8.1
aiofiles==23.2.1
python-dotenv==1.0.0
